% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_pixels_to_area_gui.R
\name{convert_pixels_to_area_gui}
\alias{convert_pixels_to_area_gui}
\title{convert pixel counts to an area based on the known area of an identifier-dot.}
\usage{
convert_pixels_to_area_gui(pixel.counts, identifier_area)
}
\arguments{
\item{pixel.counts}{list containing any of the following pixel-counts:
\itemize{
\item "bex_green_HSV"
\item "bex_drought_HSV"
\item "bex_complete_HSV"
\item "bex_identifier_HSV"
}

The pixel-count for \code{identifier} must be present, whereas the others are optional.
The calculated area is internally unit-less. However, as the default value for
\code{duflor.default_identifier_area} is 0.503 [cm^2], the areas computed by this
function is also in [cm^2].

The assumed area of the identifier can be modified by excuting
\verb{options(duflor.default_identifier_area = <value_in_square_centimeters>)} \emph{prior} to calling this function}

\item{identifier_area}{declared size of the identifier area, in \verb{[cm^2]}}
}
\value{
list of computed areas for any key listed in \code{pixel.counts}.
Additionally, two meta-values are returned as well:
\itemize{
\item the identifier-area used for calculating each area (whichever value is set in option \code{duflor.default_identifier_area})
\item the area of a singe pixel
}

All values are in [cm^2]
}
\description{
convert pixel counts to an area based on the known area of an identifier-dot.
}
\note{
This is a restructured version of \link[duflor:convert_pixels_to_area]{duflor::convert_pixels_to_area}, to be more suitable to the needs of this specific app.
}
\keyword{internal}
