% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_distortion.R
\name{correct_distortion}
\alias{correct_distortion}
\title{correct for (barrel-) distortion}
\usage{
correct_distortion(
  pixel.idx,
  distortions,
  image_dimensions,
  do_crop_image,
  x0,
  y0
)
}
\arguments{
\item{pixel.idx}{matrix of pixel-coordinate pairs for matched pixels.}

\item{distortions}{list of distortion-factors to apply. Values must be numeric, on a scale from -100 <> +100
Currently, only \code{barrel}-distortion can be applied.}

\item{image_dimensions}{actual dimensions of the image, without cropping.}

\item{do_crop_image}{boolean to check if the image was cropped.}

\item{x0}{coordinates of the cropping top left corner. Necessary to correct coordinate-space for cropping.}

\item{y0}{''}
}
\value{
sum of distortion-corrected pixels, rounded to the closest integer value.
}
\description{
correct for (barrel-) distortion
}
\keyword{internal}
