% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_key_handler.R
\name{dev_key_handler}
\alias{dev_key_handler}
\title{handles the setting and unsetting of debug-keys}
\usage{
dev_key_handler(input, DATA, DEBUGKEYS, session, use_logical_cores)
}
\arguments{
\item{input}{respective shiny-component}

\item{DATA}{respective shiny-component}

\item{DEBUGKEYS}{respective shiny-component}

\item{session}{shiny session-object}

\item{use_logical_cores}{internal flag}
}
\value{
modified \code{DEBUGKEYS}?
}
\description{
Keys which begin with 2 \code{-} are considered boolean flags, and must be provided
in the format \verb{--key=<boolean>}, where for \code{boolean} the following values are
allowed:
\itemize{
\item \code{T}/\code{TRUE}/\code{1}
\item \code{F}/\code{FALSE}/\code{0}
}
}
\details{
Keys which begin with 3 \code{-} are expecting non-boolean, custom input, which will differ
based on the respective key.
}
\keyword{internal}
