% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_args.R
\name{.main_args}
\alias{.main_args}
\title{pseudo-function to define default parameter documentation}
\usage{
.main_args(input, output, DATA, FLAGS, DEBUGKEYS, session, volumes)
}
\arguments{
\item{input}{respective shiny-component}

\item{output}{respective shiny-component}

\item{DATA}{respective shiny-component}

\item{FLAGS}{respective shiny-component}

\item{DEBUGKEYS}{respective shiny-component}

\item{session}{shiny session-object}

\item{volumes}{volumes present on the machine}
}
\value{
nothing. This function does \strong{literally nothing}
}
\description{
pseudo-function to define default parameter documentation
}
\keyword{internal}
