% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{duflor_gui}
\alias{duflor_gui}
\title{Start the duflor-gui via this function
importFrom shiny
importFrom stringr str_
importFrom stringr str_}
\usage{
duflor_gui()
}
\value{
.
}
\description{
Start the duflor-gui via this function
importFrom shiny
importFrom stringr str_
importFrom stringr str_
}
