% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group.R
\name{get_group}
\alias{get_group}
\title{trims numbers from groups from character-vector}
\usage{
get_group(names)
}
\arguments{
\item{names}{.}
}
\value{
.
}
\description{
trims numbers from groups from character-vector
}
\keyword{internal}
