% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_image_dimensions.R
\name{get_image_dimensions}
\alias{get_image_dimensions}
\title{retrieves image dimension via call to system interface}
\usage{
get_image_dimensions(path)
}
\arguments{
\item{path}{path to file to measure}
}
\value{
\code{list(width=dbl,height=dbl)}
}
\description{
retrieves image dimension via call to system interface
}
