% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_KPI_plot.R
\name{get_KPI_plot}
\alias{get_KPI_plot}
\title{generate a ggplot for a given KPI}
\usage{
get_KPI_plot(input, DATA)
}
\arguments{
\item{input}{respective shiny-component}

\item{DATA}{respective shiny-component}
}
\value{
ggplot-object for the respective KPI
}
\description{
Function generates a ggplot-object base on the selected choices in the DDLs
}
\keyword{internal}
