% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit_identifier_coordinates.R
\name{limit_identifier_coordinates}
\alias{limit_identifier_coordinates}
\title{restrict area in which \code{identifier}-pixels are considered "valid"}
\usage{
limit_identifier_coordinates(
  spectrums_object,
  image_dimensions,
  identifiersearch_x0,
  identifiersearch_x1,
  identifiersearch_y0,
  identifiersearch_y1
)
}
\arguments{
\item{spectrums_object}{return object of \code{\link[=extract_pixels_HSV]{extract_pixels_HSV()}}}

\item{image_dimensions}{dimensions of image}

\item{identifiersearch_x0}{respective coordinate of the search bounding rectangle}

\item{identifiersearch_x1}{respective coordinate of the search bounding rectangle}

\item{identifiersearch_y0}{respective coordinate of the search bounding rectangle}

\item{identifiersearch_y1}{respective coordinate of the search bounding rectangle}
}
\value{
modified \code{spectrums_object}
}
\description{
This function is only applied to the \code{identifier}-range.
The intention is to exclude random pixels which \strong{definitely} should not be
attributed to the identifier-data.
}
\keyword{internal}
