% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit_to_range.R
\name{limit_to_range}
\alias{limit_to_range}
\title{replace values outside of boundaries with respective  boundary}
\usage{
limit_to_range(x, replace_lower, replace_upper)
}
\arguments{
\item{replace_lower}{value to insert in place of elements of \code{value} below this bound.}

\item{replace_upper}{see \code{replace_lower}}

\item{value}{vector to normalise}
}
\value{
vector, normalised to range \verb{[replace_lower, replace_upper]}
}
\description{
This function limits only values which lie outside of
\verb{[replace_lower, replace_upper}]. Values which do not fall outside this range
are not modified.
}
\examples{
result <- duflor.gui:::limit_to_range(c(0,100,255),0,255)
result2 <- duflor.gui:::limit_to_range(c(0,100,256),0,255)
print(result)
print(result2)
}
\keyword{internal}
