% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_parallelPanel_event.R
\name{open_parallelPanel_event}
\alias{open_parallelPanel_event}
\title{internal callback function for the checkbox opening the parallelisation-panel}
\usage{
open_parallelPanel_event(
  input,
  DATA,
  FLAGS,
  use_logical_cores,
  session,
  STARTUP,
  default_init_cores = 2
)
}
\arguments{
\item{input}{respective shiny-component}

\item{DATA}{respective shiny-component}

\item{FLAGS}{respective shiny-component}

\item{use_logical_cores}{internal flag}

\item{session}{shiny session-object}

\item{STARTUP}{respective shiny-component}
}
\description{
internal callback function for the checkbox opening the parallelisation-panel
}
\keyword{internal}
