% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_resultsObject.R
\name{prepare_resultsObject}
\alias{prepare_resultsObject}
\title{prepare the dataframe-object into which all results are to be merged.}
\usage{
prepare_resultsObject(input, DATA, DEBUGKEYS)
}
\arguments{
\item{input}{respective shiny-component}

\item{DATA}{respective shiny-component}

\item{DEBUGKEYS}{respective shiny-component}
}
\value{
results-csv-framework.
}
\description{
results requires fields
\itemize{
\item full_path
\item image_date
\item image_name
\item processed_width
\item processed_height
\item for each in  \code{names(DATA$spectrums$lower_bound)}:
\itemize{
\item \%spectrum\%_count
\item img.fraction (maybe not that valuable?)
\item \%spectrum\%_area (as results of duflor::convert_pixels_to_area())
\item \%spectrum\%_fraction (as defined as \verb{\%spectrum\%_count/(sum(<all_plant_spectrums>_count)})
}
}
}
\keyword{internal}
