% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_selected_mask.R
\name{render_selected_mask}
\alias{render_selected_mask}
\title{render mask based on input data}
\usage{
render_selected_mask(input, DATA)
}
\arguments{
\item{input}{\itemize{
\item \code{input} respective shiny-component
}}

\item{DATA}{\itemize{
\item \code{DATA} respective shiny-component
}}
}
\description{
render mask based on input data
}
\keyword{internal}
