% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_state.R
\name{restore_state}
\alias{restore_state}
\title{restore application state from stored \code{.RDS}-file}
\usage{
restore_state(
  input,
  output,
  DATA,
  FLAGS,
  DEBUGKEYS,
  session,
  volumes,
  state_file
)
}
\arguments{
\item{input}{\itemize{
\item \code{input} respective shiny-component
}}

\item{output}{\itemize{
\item \code{output} respective shiny-component
}}

\item{DATA}{\itemize{
\item \code{DATA} respective shiny-component
}}

\item{FLAGS}{\itemize{
\item \code{FLAGS} respective shiny-component
}}

\item{DEBUGKEYS}{\itemize{
\item shiny-app \code{DEBUGKEYS}-reactives
}}

\item{session}{\itemize{
\item shiny session
}}

\item{volumes}{\itemize{
\item volume letters available
}}

\item{state_file}{filepath for to-be-loaded state_file}
}
\value{
path to the loaded file.
}
\description{
This function is responsible for
}
\keyword{internal}
