% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_error_state.R
\name{save_error_state}
\alias{save_error_state}
\title{Handle saving application-state to file}
\usage{
save_error_state(
  input,
  DATA,
  DEBUGKEYS,
  FLAGS,
  volumes,
  error,
  errordir_path = NA,
  erroneous_callback = ""
)
}
\arguments{
\item{input}{\itemize{
\item \code{input} respective shiny-component
}}

\item{DATA}{\itemize{
\item \code{DATA} respective shiny-component
}}

\item{DEBUGKEYS}{\itemize{
\item shiny-app \code{DEBUGKEYS}-reactives
}}

\item{volumes}{\itemize{
\item volume letters available
}}

\item{error}{error-object}

\item{errordir_path}{directory to write error-replication state to}
}
\value{
path to error-state file which may be used to load the error-triggering settings.
}
\description{
\itemize{
\item assumes \code{input$save_state} to be properly formatted, as returned by \code{\link[shinyFiles:shinyFiles-observers]{shinyFiles::shinyFileSave()}}
\item assumes \code{volumes} to be the return-value of \code{\link[shinyFiles:getVolumes]{shinyFiles::getVolumes()}}
}
}
\keyword{internal}
