% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_state.R
\name{save_state}
\alias{save_state}
\title{Title}
\usage{
save_state(input, DATA, DEBUGKEYS, FLAGS, volumes)
}
\arguments{
\item{input}{.}

\item{DATA}{.}

\item{DEBUGKEYS}{.}

\item{FLAGS}{.}

\item{volumes}{.}
}
\value{
.
}
\description{
Title
}
\keyword{internal}
