% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_spectra_gui_comp.R
\name{select_spectra_gui_comp}
\alias{select_spectra_gui_comp}
\title{generate and call a modal-dialogue which lets you select which spectrums to analyse.}
\usage{
select_spectra_gui_comp(input)
}
\arguments{
\item{input}{\code{input}-shiny-component}
}
\value{
.
}
\description{
Additionally, it queries the component \code{radio_analysis_type} to preselect common results for different types of analyses
}
\keyword{internal}
