% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_results_to_file.R
\name{store_results_to_file}
\alias{store_results_to_file}
\title{write results of the analysis to file}
\usage{
store_results_to_file(results, results_path, save_to_xlsx = FALSE)
}
\arguments{
\item{results}{object to write.}

\item{results_path}{path to write file to, sans the filesuffix.}

\item{save_to_xlsx}{should results be written into a \code{.xlsx}-file? If false, it will be written into a \code{.csv}-file instead.}
}
\value{
list containing
\itemize{
\item boolean check whether or not file was written to disk successfully
\item complete path to the resulting file.
}
}
\description{
write results of the analysis to file
}
\note{
\code{XLSX}-files are the suggested output if any additional postprocessing is required.
If the data will not be modified again, writing it to \code{csv} will save space.
}
\keyword{internal}
