% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_resultsObject.R
\name{update_resultsObject}
\alias{update_resultsObject}
\title{adds a line of values \code{newline} to a df \code{resultsObject}}
\usage{
update_resultsObject(resultsObject, new_line)
}
\arguments{
\item{resultsObject}{existing dataframe}

\item{new_line}{data-vector to add}
}
\value{
updated \code{resultsObject}
}
\description{
This function does not validate its inputs!.
}
\keyword{internal}
