% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_custom_HSV_values.R
\name{validate_custom_HSV_values}
\alias{validate_custom_HSV_values}
\title{ensure that custom HSV values are within required range}
\usage{
validate_custom_HSV_values(input, DATA, session)
}
\arguments{
\item{input}{respective shiny-component}

\item{DATA}{respective shiny-component}

\item{session}{shiny session-object}
}
\value{
list containing
\itemize{
\item expected bounds
\item keys for which a specific bound was violated, along its violated value
}
}
\description{
For \code{HUE}, value must be within \verb{[0, 359]}
For \code{SATURATION}, value must be within \verb{[0, 1]}
For \code{VALUE}, value must be within \verb{[0, 1]}
}
\keyword{internal}
