% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_analysis.R
\name{execute_analysis}
\alias{execute_analysis}
\title{wrapper-function facilitating single- and multi-image analysis.}
\usage{
execute_analysis(input, DATA, DEBUGKEYS, FLAGS)
}
\arguments{
\item{input}{\itemize{
\item \code{input} respective shiny-component
}}

\item{DATA}{\itemize{
\item \code{DATA} respective shiny-component
}}

\item{DEBUGKEYS}{\itemize{
\item \code{DEBUGKEYS} respective shiny-component
}}

\item{FLAGS}{\itemize{
\item \code{FLAGS} respective shiny-component
}}
}
\value{
list with components
\itemize{
\item \code{results}: results-data passed-through from \code{\link[=execute_multiple]{execute_multiple()}} or \code{\link[=execute_single]{execute_single()}}
}
}
\description{
wrapper-function facilitating single- and multi-image analysis.
}
