% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_key_handler.R
\name{key_handle_cores}
\alias{key_handle_cores}
\title{internal function handling dev-key \code{---set.cores=XX}}
\usage{
key_handle_cores(value, DATA, use_logical_cores, session)
}
\arguments{
\item{value}{number of cores to assign}

\item{DATA}{respective shiny-component}

\item{use_logical_cores}{whether or not to use logical cores}

\item{session}{shiny session-object}
}
\description{
internal function handling dev-key \code{---set.cores=XX}
}
\keyword{internal}
