% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_parallelPanel_event.R
\name{open_parallelPanel_event}
\alias{open_parallelPanel_event}
\title{internal callback function for the checkbox opening the parallelisation-panel}
\usage{
open_parallelPanel_event(
  input,
  DATA,
  use_logical_cores,
  session,
  STARTUP,
  default_init_cores = 2
)
}
\arguments{
\item{input}{\itemize{
\item \code{input} respective shiny-component
}}

\item{DATA}{\itemize{
\item \code{DATA} respective shiny-component
}}

\item{use_logical_cores}{\itemize{
\item internal flag
}}

\item{session}{\itemize{
\item shiny session
}}

\item{STARTUP}{\itemize{
\item \code{STARTUP} respective shiny-component
}}
}
\description{
internal callback function for the checkbox opening the parallelisation-panel
}
\keyword{internal}
