% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_KPI_plot_to_file.R
\name{store_KPI_plot_to_file}
\alias{store_KPI_plot_to_file}
\title{save KPI-plot to file}
\usage{
store_KPI_plot_to_file(input, DATA)
}
\arguments{
\item{input}{\itemize{
\item \code{input} respective shiny-component
}}

\item{DATA}{\itemize{
\item \code{DATA} respective shiny-component
}}
}
\description{
save KPI-plot to file
}
\keyword{internal}
